fix warnings

+++ src/arch/festival/server.cc
@@ -241,7 +241,7 @@ static int client_access_check(int fd,int client)
     if ((client_access == TRUE) && (passwd != NULL))
     {
 	char *client_passwd = walloc(char,strlen(get_c_string(passwd))+1);
-	if (read(fd,client_passwd,strlen(get_c_string(passwd))) !=
+	if ((unsigned int)read(fd,client_passwd,strlen(get_c_string(passwd))) !=
             (unsigned int)strlen(get_c_string(passwd)))
         {
             client_access = FALSE;
